revise Edge; revise EdgeWeightedGraph;

+++ src/MinimumSpanningTrees/Edge.java
@@ -4,7 +4,7 @@ public class Edge implements Comparable<Edge> {
     private final int v;
     private final int w;
     private final double weight;
-    public Edge(int v, int w, int weight)//构造函数
+    public Edge(int v, int w, double weight)//构造函数
     {
         this.v= v;
         this.w= w;
